Fix mem_map/munmap/mprotect bugs in memory management
- munmap: throw IllegalStateException when adjacent region exceeds remaining unmap size (prevents size going negative)
- findMemoryMap: fix validation to precisely check unmap range against segment bounds
- Remove unnecessary (int) casts on MemoryMap size fields (MemoryMap.size is long)
- mprotect: align length to page boundary before calling backend

Made-with: Cursor

+++ unidbg-android/src/main/java/com/github/unidbg/linux/AndroidElfLoader.java
@@ -408,7 +408,7 @@ public class AndroidElfLoader extends AbstractLoader<AndroidFileIO> implements M
                                 if (mMapListener != null) {
                                     mMapListener.onMap(base, off, UnicornConst.UC_PROT_NONE);
                                 }
-                                if (memoryMap.put(base, new MemoryMap(base, (int) off, UnicornConst.UC_PROT_NONE)) != null) {
+                                if (memoryMap.put(base, new MemoryMap(base, off, UnicornConst.UC_PROT_NONE)) != null) {
                                     log.warn("mem_map replace exists memory map base={}", Long.toHexString(base));
                                 }
                             }
